+++ agent-default/setup.go
@@ -12,8 +12,8 @@ import (
 )
 
 const (
-	setupOperationName  = "agent.default.config"
-	setupOperationGroup = "configuration"
+	setupOperationName  = "config"
+	setupOperationGroup = "agent-default"
 )
 
 // setupOperation asks the Host for this Plugin's configuration through a